+++ packages/fhir-validation/src/utilities.ts
@@ -1,3 +1,29 @@
+import { primitiveTypes, resourceTypes } from "@iguhealth/fhir-types/r4/sets";
+import { ElementDefinition } from "@iguhealth/fhir-types/r4/types";
+
 export function validateIsObject(v: unknown): v is object {
   return typeof v === "object" && v !== null;
+}
+
+export function capitalize(str: string) {
+  return str.charAt(0).toUpperCase() + str.slice(1);
+}
+
+export function isPrimitiveType(type: string) {
+  return primitiveTypes.has(type);
+}
+
+export function notNull<T, Z extends T | undefined>(
+  element: Z,
+): element is NonNullable<Z> {
+  if (element === undefined || element === null) return false;
+  return true;
+}
+
+export function isTypeChoice(element: ElementDefinition) {
+  return (element.type || []).length > 1;
+}
+
+export function isResourceType(type: string) {
+  return resourceTypes.has(type);
 }
